# openapi.yaml
openapi: 3.0.1
info:
  title: "Multipass Plugin API"
  version: "0.0.1"
paths:
  /create_image:
    post:
      summary: "Create an image from text"
      operationId: "createImage"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                prompt:
                  type: "string"
      responses:
        200:
          description: "Image created"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  status:
                    type: "string"
                  result:
                    type: "string"
  /create_image_from_drawing:
    post:
      summary: "Create an image from a drawing or another image"
      operationId: "createImageFromDrawing"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                prompt:
                  type: "string"
                image_url:
                  type: "string"
      responses:
        200:
          description: "Image created"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  status:
                    type: "string"
                  result:
                    type: "string"
  /create_video:
    post:
      summary: "Create a video from text"
      operationId: "createVideo"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              properties:
                prompt:
                  type: "string"
      responses:
        200:
          description: "Image created"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  status:
                    type: "string"
                  url:
                    type: "string"